requirements-agents.txt
# Agents IA
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
langchain>=0.0.300
llama-index>=0.8.0

# Web3 & Blockchain
web3>=6.0.0
eth-account>=0.9.0
py-solc-x>=2.0.0

# Development Tools
watchdog>=3.0.0
click>=8.0.0
rich>=13.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Security & Audit (optionnels - nécessitent Build Tools Windows)
# mythril>=0.23.0
# slither-analyzer>=0.9.0

# Utilities
requests>=2.31.0
aiohttp>=3.9.0
tenacity>=8.2.0